requirements.txt
fastapi
uvicorn
websockets
pyyaml
numpy
opencv-python
ndi-python